backend/requirements.txt
fastapi==0.95.2
uvicorn==0.24.0
python-dotenv==1.0.0
tortoise-orm==0.19.3
asyncpg==0.29.0
redis==5.0.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
pydantic==1.10.13
aerich==0.7.2 